TO BE USED FOR EDUCATIONAL PURPOSES ONLY
The use of the d0m1_darkddos & its resources is COMPLETE RESPONSIBILITY of the END-USER. Developers assume NO liability and are NOT responsible for any misuse or damage caused by this program. 
Also we inform you that some of your your actions may be ILLEGAL and you CAN NOT use this software to test person or company without WRITTEN PERMISSION from them.